Update review_status && move FE service into services
