+++ client-react/src/components/login/Form.jsx
@@ -14,6 +14,21 @@ import { setMyProfile } from "../../redux/myProfileSlice";
 
 // todo : 之后可能改用FormBox。
 // 该组件之后会进行修改，通过判断渲染成register或者是login
+const checkReviewStatus = (profile) => {
+    let reviewStatus = profile?.review_status || "Never Submitted";
+    if (reviewStatus === true) {
+        reviewStatus = 'Approved';
+    } else if (reviewStatus === false) {
+        reviewStatus = 'Rejected';
+    } else {
+        reviewStatus = 'Pending';
+    }
+
+    return {
+        ...profile,
+        review_status: reviewStatus
+    };
+}
 
 export default function Form() {
     const [showPassword, setShowPassword] = React.useState(false);
@@ -36,6 +51,7 @@ export default function Form() {
 
                 if (response.status === 201) {
                     console.log("Login Success", response);
+                    response._doc.personal_info = checkReviewStatus(response._doc.personal_info);
                     dispatch(setIsLogin(true));
                     dispatch(
                         setUser({
